`unsigned`. `unsigned` everywhere.

+++ quark/api/quark_plugin_version.h
@@ -9,9 +9,9 @@
 #define quark_plugin_version_h
 
 struct QuarkPluginVersion {
-    const int major;
-    const int minor;
-    const int patch;
+    const unsigned short major;
+    const unsigned short minor;
+    const unsigned short patch;
 };
 
 #endif /* quark_plugin_version_h */
